python_dc/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
fuzzywuzzy
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
cryptography==41.0.7
dnspython==2.5.0
fastapi==0.109.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.2
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.2
idna==3.6
inexactsearch==1.0.2
Jinja2==3.1.3
joblib==1.3.2
keras==2.15.0
langdetect==1.0.9
language-detector==5.0.2
libclang==16.0.6
libretranslatepy==2.1.1
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyLanguagetool==0.10.0
pymongo==4.6.1
pyphen==0.14.0
PySocks==1.7.1
pyspellchecker==0.8.1
pytesseract==0.3.10
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.4.0
scipy==1.12.0
selenium==4.16.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
silpa-common==0.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.5
spellchecker==0.4
starlette==0.35.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
textblob==0.17.1
textstat==0.7.3
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
translate==3.6.1
trio==0.24.0
trio-websocket==0.11.1
triton==2.1.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
webdriver-manager==4.0.1
Werkzeug==3.0.1
whois==0.99.3
wrapt==1.14.1
wsproto==1.2.0
